backend/requirements.txt
pdfplumber
pymupdf
sentence-transformers
faiss-cpu
numpy
transformers
accelerate
flask
flask-cors
